requirements/test.txt
-r common.txt
django-storages==1.1.5
django-compressor==1.2
django-extensions==1.0.0
boto==2.6.0


